Updated endpoints with auth

+++ index.js
@@ -27,6 +27,12 @@ app.use(bodyParser.urlencoded({
 
 app.use(bodyParser.json());
 app.use(methodOverride());
+
+//Authentication import
+let auth = require('./auth.js')(app);
+const passport = require('passport');
+require('./passport.js');
+
 //setup Logger
 app.use(morgan('combined', {stream: accessLogStream}));
 app.use(express.static('public')); //routes all requests for static files to their reapective files on the server
@@ -186,7 +192,7 @@ app.get('/documentation.html', (req,res) => {
 
 //MOVIE ENDPOINTS
 //READ ALL movies
-app.get('/movies', async (req, res) => {
+app.get('/movies', passport.authenticate('jwt', {session: false}), async (req, res) => {
     await Movies.find()
         .then((movies) => {
             res.status(201).json(movies);
@@ -198,7 +204,7 @@ app.get('/movies', async (req, res) => {
 });
 
 //READ description of movie by title
-app.get('/movies/:Title', async (req, res) => {
+app.get('/movies/:Title', passport.authenticate('jwt', {session: false}), async (req, res) => {
     await Movies.findOne({Title: req.params.Title})
         .then((movie) => {
             res.json(movie);
@@ -210,7 +216,7 @@ app.get('/movies/:Title', async (req, res) => {
 });
 
 //READ all genres
-app.get('/genres', async (req,res) => {
+app.get('/genres', passport.authenticate('jwt', {session: false}), async (req,res) => {
     await Genres.find()
         .then((genres) => {
             res.json(genres);
@@ -222,7 +228,7 @@ app.get('/genres', async (req,res) => {
 });
 
 //GET movies by genre
-app.get('/movies/Genre/:Name', async (req, res) => {
+app.get('/movies/Genre/:Name', passport.authenticate('jwt', {session: false}), async (req, res) => {
     await Movies.find({"Genre.Name": req.params.Name})
         .then((movieGenres) => {
             res.json(movieGenres);
@@ -234,7 +240,7 @@ app.get('/movies/Genre/:Name', async (req, res) => {
 });
 
 //READ data about a dircetor by name
-app.get('/movies/director/:Name', async (req, res) => {
+app.get('/movies/director/:Name', passport.authenticate('jwt', {session: false}), async (req, res) => {
     await Movies.findOne({"Director.Name": req.params.Name})
         .then((director) => {
             res.json(director);
@@ -282,7 +288,12 @@ app.post('/users', async (req, res) =>{
 });
 
 //READ all users
-app.get('/users', async (req, res) => {
+app.get('/users', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    //Condition Check
+    if (req.user.Username != req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+    //End Condition
     await Users.find()
     .then((users) => {
         res.status(201).json(users);
@@ -294,7 +305,12 @@ app.get('/users', async (req, res) => {
 });
 
 //GET a user by username
-app.get('/users/:Username', async(req, res) => {
+app.get('/users/:Username', passport.authenticate('jwt', {session: false}), async(req, res) => {
+    //Condition Check
+    if (req.user.Username != req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+    //End Condition
     await Users.findOne({Username: req.params.Username})
 
     .then((user) => {
@@ -313,7 +329,12 @@ app.get('/users/:Username', async(req, res) => {
     Email: String, (required)
     Birthday: Date
 } */
-app.put('/users/:Username', async (req, res) => {
+app.put('/users/:Username', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    //Condition Check
+    if (req.user.Username != req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+    //End Condition
     await Users.findOneAndUpdate({Username: req.params.Username}, {$set: 
     {
         Username: req.body.Username,
@@ -359,7 +380,12 @@ app.put('/users/:Username', async (req, res) => {
 
 
 //UPDATE movie to user's favoriteMovies. Consider using "$addToSet" instead of "$push"
-app.post('/users/:Username/movies/:MovieID', async (req, res) => {
+app.post('/users/:Username/movies/:MovieID', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    //Condition Check
+    if (req.user.Username != req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+    //End Condition
     await Users.findOneAndUpdate({Username: req.params.Username}, {
         $push: {FavoriteMovies: req.params.MovieID}
     },
@@ -374,7 +400,12 @@ app.post('/users/:Username/movies/:MovieID', async (req, res) => {
 });
 
 //DELETE movie from user favoriteList. Same as add, but use $pull
-app.delete('/users/:Username/movies/:MovieID', async (req,res) => {
+app.delete('/users/:Username/movies/:MovieID', passport.authenticate('jwt', {session: false}), async (req,res) => {
+   //Condition Check
+   if (req.user.Username != req.params.Username) {
+    return res.status(400).send('Permission denied');
+}
+//End Condition
     await Users.findOneAndUpdate({Username: req.params.Username}, {
         $pull: {FavoriteMovies: req.params.MovieID}
     },
@@ -389,7 +420,12 @@ app.delete('/users/:Username/movies/:MovieID', async (req,res) => {
 });
 
 //DELETE user, return text that user email has been removed
-app.delete('/users/:Username', async (req, res) => {
+app.delete('/users/:Username', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    //Condition Check
+    if (req.user.Username != req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+    //End Condition
     await Users.findOneAndRemove({Username: req.params.Username})
         .then((user) => {
             if (!user) {
